requirements-airflow.txt
duckdb
kafka-python
apache-airflow-providers-apache-kafka>=1.9.0
apache-airflow-providers-common-messaging>=1.0.3
confluent-kafka
